+++ pkg/apis/replication/v2alpha1/replication_status.go
@@ -36,6 +36,7 @@ type DeploymentReplicationStatus struct {
 	// Destination contains the detailed status of the destination endpoint
 	Destination EndpointStatus `json:"destination"`
 
+	// Deprecated: this field will not be updated anymore
 	// CancelFailures records the number of times that the configuration was canceled
 	// which resulted in an error.
 	CancelFailures int `json:"cancel-failures,omitempty"`